+++ RELEASE_NOTES.rst
@@ -7,6 +7,7 @@ Release notes
 * Enhance ``doc-tools-check-languages`` to handle translation of RST
   guides and publishing of draft guides to /draft/.
 * ``autohelp.py``: lookup configuration options in more oslo libraries.
+* ``autohelp.py``: add a hook for neutron plugins
 * ``autohelp-wrapper``: improve reliability by building a virtual env per
   project, rather than a common virtual env.
 * ``autohelp-wrapper``: define the custom dependencies for each project in